+++ Template/Tests/Usings.cs
@@ -1,0 +1,7 @@
+﻿global using FastEndpoints;
+global using FastEndpoints.Testing;
+global using FluentAssertions;
+global using MongoDB.Entities;
+global using Xunit;
+global using Xunit.Abstractions;
+global using Xunit.Priority;